// apps/service-web/src/.generated/api/model/eventItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Service API
 * Data Plane Service API Service
 * OpenAPI spec version: 1.0.0
 */
import type { EventItemStatus } from './eventItemStatus';

export interface EventItem {
  id: string;
  title: string;
  content: string;
  startsAt: string;
  endsAt: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  linkUrl: string | null;
  status: EventItemStatus;
  /** @nullable */
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}