// apps/service-web/src/.generated/api/model/eventsControllerListV1Phase.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Service API
 * Data Plane Service API Service
 * OpenAPI spec version: 1.0.0
 */

export type EventsControllerListV1Phase = typeof EventsControllerListV1Phase[keyof typeof EventsControllerListV1Phase];


export const EventsControllerListV1Phase = {
  ongoing: 'ongoing',
  upcoming: 'upcoming',
  ended: 'ended',
} as const;